requirements.txt
anything==0.2.1
certifi==2018.4.16
chardet==3.0.4
clarifai==2.2.3
configparser==3.5.0
dj-database-url==0.5.0
Django==2.0.4
django-appconf==1.0.2
django-classy-tags==0.8.0
django-extensions==2.1.0
django-heroku==0.3.1
django-images==0.4.3
django-markdownx==2.0.23
django-tables2==1.21.2
django-taggit==0.23.0
django-taggit-templatetags2==1.6.1
django-tinymce==2.7.0
django-tinymce4==3.0.2
django-tinymce4-lite==1.7.2
easy-thumbnails==2.5
EasyProcess==0.2.3
emoji==0.5.0
future==0.16.0
gunicorn==19.9.0
idna==2.6
jsmin==2.2.2
jsonschema==2.6.0
Markdown==2.6.11
pilkit==2.0
Pillow==5.1.0
psycopg2==2.7.5
pydub==0.22.1
python-interface==1.4.0
pytz==2018.4
PyVirtualDisplay==0.2.1
requests==2.18.4
requests-toolbelt==0.8.0
selenium==3.12.0
six==1.11.0
urllib3==1.22
whitenoise==4.0
